backend/requirements.txt
asgiref==3.9.1
Django==4.2.23
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
pytz==2025.2
sqlparse==0.5.3
typing_extensions==4.14.1
python-dotenv==1.0.0
Pillow==10.2.0
